dont just remove the source file if src and dest are the same

+++ deltachat-ios/Helper/FileHelper.swift
@@ -89,6 +89,7 @@ public class FileHelper {
     }
 
     static func copyIfPossible(src: URL, dest: URL) -> URL {
+        guard src != dest else { return }
         do {
             if FileManager.default.fileExists(atPath: dest.path) {
                 try FileManager.default.removeItem(at: dest)
